refactor(migrates): add some abstract methods for migrates
