+++ server/odc-migrate/src/main/java/com/oceanbase/odc/migrate/AbstractMetaDBMigrate.java
@@ -35,6 +35,7 @@ import org.springframework.integration.jdbc.lock.JdbcLockRegistry;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import com.oceanbase.odc.common.lang.Holder;
+import com.oceanbase.odc.core.migrate.BootstrapSchemaHistoryRepository;
 import com.oceanbase.odc.core.migrate.MigrateConfiguration;
 import com.oceanbase.odc.core.migrate.Migrates;
 
@@ -71,7 +72,8 @@ abstract public class AbstractMetaDBMigrate {
                 log.info("init configuration success, migrate starting, initVersion={}",
                         configuration.getInitVersion());
 
-                new Migrates(configuration).migrate();
+                new Migrates(configuration, new BootstrapSchemaHistoryRepository(
+                        configuration.getDataSource())).migrate();
                 log.info("migrate success");
             } finally {
                 lock.unlock();
